+++ pkg/plugin.go
@@ -5,22 +5,26 @@ import (
 	"encoding/json"
 	"net/http"
 
+	"github.com/gorilla/mux"
 	"github.com/grafana/grafana-plugin-sdk-go/backend"
 	"github.com/grafana/grafana-plugin-sdk-go/backend/datasource"
 	"github.com/grafana/grafana-plugin-sdk-go/backend/instancemgmt"
+	"github.com/grafana/grafana-plugin-sdk-go/backend/resource/httpadapter"
 	"github.com/grafana/grafana-plugin-sdk-go/data"
+	"github.com/yesoreyeram/grafana-infinity-datasource/pkg/infinity"
 )
 
-type queryModel struct{}
-
 func newDatasource() datasource.ServeOpts {
-	im := datasource.NewInstanceManager(newDataSourceInstance)
-	ds := &InfinityDatasource{
-		im: im,
+
+	handler := &InfinityDatasource{
+		im: datasource.NewInstanceManager(newDataSourceInstance),
 	}
+	router := mux.NewRouter()
+	router.HandleFunc("/proxy", handler.proxyHandler)
 	return datasource.ServeOpts{
-		QueryDataHandler:   ds,
-		CheckHealthHandler: ds,
+		QueryDataHandler:    handler,
+		CheckHealthHandler:  handler,
+		CallResourceHandler: httpadapter.New(router),
 	}
 }
 
@@ -48,25 +52,45 @@ func (ds *InfinityDatasource) QueryData(ctx context.Context, req *backend.QueryD
 }
 
 func (ds *InfinityDatasource) query(ctx context.Context, query backend.DataQuery) (response backend.DataResponse) {
-	var qm queryModel
+	var qm infinity.Query
 	response.Error = json.Unmarshal(query.JSON, &qm)
 	if response.Error != nil {
 		return response
 	}
 	frame := data.NewFrame("response")
 	response.Frames = append(response.Frames, frame)
-
 	return response
 }
 
 type instanceSettings struct {
-	httpClient *http.Client
+	client *infinity.Client
 }
 
 func (is *instanceSettings) Dispose() {}
 
 func newDataSourceInstance(setting backend.DataSourceInstanceSettings) (instancemgmt.Instance, error) {
+	settings, err := infinity.LoadSettings(setting)
+	if err != nil {
+		return nil, err
+	}
+	client, err := infinity.NewClient(settings)
+	if err != nil {
+		return nil, err
+	}
 	return &instanceSettings{
-		httpClient: &http.Client{},
+		client: client,
 	}, nil
+}
+
+func getInstance(im instancemgmt.InstanceManager, ctx backend.PluginContext) (*instanceSettings, error) {
+	instance, err := im.Get(ctx)
+	if err != nil {
+		return nil, err
+	}
+	return instance.(*instanceSettings), nil
+}
+
+func getInstanceFromRequest(im instancemgmt.InstanceManager, req *http.Request) (*instanceSettings, error) {
+	ctx := httpadapter.PluginConfigFromContext(req.Context())
+	return getInstance(im, ctx)
 }